const, utils - added files
