+++ src/const.js
@@ -1,0 +1,12 @@
+export const EVENT_TYPE = [
+  `taxi`,
+  `bus`,
+  `train`,
+  `ship`,
+  `transport`,
+  `drive`,
+  `flight`,
+  // `checkIn`,
+  `sightseeing`,
+  `restaurant`,
+];
